Title: 쪼개어 배낭 채우기 구현, Time: 19ms, Memory: 1MB, Status: Passed - Codetree

+++ 260314/쪼개어 배낭 채우기 구현/implement-fractional-knapsack.cpp
@@ -36,7 +36,7 @@ int main() {
         }
         else
         {
-            ans += (double)ratio * M;
+            ans += ratio * M;
             break;
         }
     }
